refactor : 서브쿼리 도메인 별로 클래스 분리
