+++ src/main/java/app/bottlenote/picks/repository/PicksQuerySupporter.java
@@ -1,5 +1,6 @@
 package app.bottlenote.picks.repository;
 
+import app.bottlenote.picks.constant.PicksStatus;
 import com.querydsl.core.types.Expression;
 import com.querydsl.core.types.ExpressionUtils;
 import com.querydsl.core.types.dsl.BooleanExpression;
@@ -10,9 +11,26 @@ import org.springframework.stereotype.Component;
 
 import static app.bottlenote.alcohols.domain.QAlcohol.alcohol;
 import static app.bottlenote.picks.domain.QPicks.picks;
+import static com.querydsl.jpa.JPAExpressions.select;
 
 @Component
 public class PicksQuerySupporter {
+
+	/**
+	 * 마이 페이지 사용자의 찜하기 개수를 조회한다.
+	 *
+	 * @param userId 마이 페이지 사용자
+	 * @return 찜하기 개수
+	 */
+	public Expression<Long> picksCountSubQuery(NumberPath<Long> userId) {
+		return ExpressionUtils.as(
+				select(picks.count())
+						.from(picks)
+						.where(picks.userId.eq(userId)
+								.and(picks.status.eq(PicksStatus.PICK))),
+				"picksCount"
+		);
+	}
 
 	/**
 	 * 사용자가 술을 찜 했는지 여부를 확인하는 서브 쿼리
